requirements.txt
# =============================================================
# RoC Desk — Python Dependencies
# =============================================================

# Django core
Django>=5.1,<5.2
psycopg2-binary>=2.9

# Environment & config
django-environ>=0.11

# Async task queue
celery>=5.4
redis>=5.0

# Media handling
Pillow>=10.4

# HTTP client (Evolution API)
requests>=2.32
qrcode>=7.4

# Static file serving
whitenoise>=6.7

# Production WSGI server
gunicorn>=23.0

# Excel export
openpyxl>=3.1

# Admin Theme
django-unfold
dnspython

# Security
django-encrypted-model-fields>=0.6
python-magic>=0.4
django-ipware>=7.0